feat: extract common expression evaluation logic for EVM and Stellar filters
