+++ src/services/filter/filters/mod.rs
@@ -6,6 +6,8 @@
 //! - EVM-specific implementation
 //! - Stellar-specific implementation
 
+mod expression;
+
 pub mod evm {
 	pub mod filter;
 	pub mod helpers;
